SMSLogs.cs: added a partial class to create entities.
HomeController.cs: added ActionResult AdminPage, involving all the information viewing the page with related to the sms data in the tables.

AdminPage.cshtml: Changed the data in the tables to be dynamic and get the data from Microsoft SQL Server Management Studio. Also fixed a bug with regards to switching from the Daily to Monthly tab (kept the Daily table).