Added Stop Name, Stop Times, and a Icon to the map markers - uses a info window.
Changed the app name to MajicBus
Changed the app icon to the icon in the project files
Added code to catch the exception that can be thrown by the JSON parsing on the map activity.

Lots of files are added in this, most of them are just generated images from adding them to the project.