added queries to site and app just needs to parse the json query now
added query to site to search database by stopid and return the route
numbers and next 5 stop times (and the sender phone number) back to the
sms app. the sms app now just needs to parse the json object for it's
data so the photo properly sends back the message.